+++ record.h
@@ -9,10 +9,10 @@ private:
     int userId;
     string date;
     string category;
-    double amount;
+    float amount;
 
 public:
-    Record (int = 0, int =0, string = "2000-01-01", string = "Unknown", double = 0.00);
+    Record (int = 0, int =0, string = "2000-01-01", string = "Unknown", float = 0.00);
     void setId(int);
     int getId();
     void setUserId(int);
@@ -21,6 +21,6 @@ public:
     string getDate();
     void setCategory(string);
     string getCategory();
-    void setAmount(double);
-    double getAmount();
+    void setAmount(float);
+    float getAmount();
 };
